requirements.txt
# Core dependencies
streamlit>=1.28.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
langchain-text-splitters>=0.0.1
chromadb>=0.4.15
pydantic-settings>=2.0.0

# PDF processing
pypdf>=3.17.0

# Embeddings
sentence-transformers>=2.2.2
openai>=1.3.0

# Configuration
python-dotenv>=1.0.0

# Utilities
tiktoken>=0.5.1
